// WYNKWeb/src/app/Models/Donor.ts
export class Donor {
  CmpID: number;
  DonorType: string;
  Name: string;
  MiddleName: string;
  LastName: string;
  Address1: string;
  Address2: string;
  Address3: string;
  Location: number;
  GSTNo: string;
  MobileNo: string;
  PhoneNo: string;
  ContactPerson: string;
  IsDeleted: boolean;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
}